discord-bot/requirements.txt
discord.py==2.3.2
google-cloud-storage==2.18.2
google-cloud-pubsub==2.26.1
